pgmdump: Variable number of fragment/vertex shaders
In the a205 GL_Z400_BINARY_AMD binary shaders, there are five
vertex shaders, not just three.

I've guessed at some header fields that might be the number
of respective shaders to avoid guessing.

This change works for me, but might potentially make the pgmdump program
incompatible with other versions of the driver, so this needs
testing first.

Signed-off-by: Wladimir J. van der Laan <laanwj@gmail.com>